+++ extensions/coc-vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # coc-vue
 
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies [f5ebe8e]
+  - @vuedx/typescript-plugin-vue@0.7.2
+
 ## 0.7.1
 
 ### Patch Changes