+++ extensions/vscode-vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # vue
 
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies [f5ebe8e]
+  - @vuedx/vue-virtual-textdocument@0.7.2
+
 ## 0.7.1
 
 ### Patch Changes